separate App.tsx into Home and Workouts pages
