+++ src/App.tsx
@@ -1,6 +1,7 @@
 import "./App.css";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Home from "./Pages/Home";
+import Workouts from "./Pages/Workouts";
 
 function App() {
   
@@ -9,6 +10,7 @@ function App() {
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<Home/>} />
+        <Route path="/workouts" element={<Workouts/>} />
       </Routes>
     </BrowserRouter>
   );
