+++ src/Pages/Home.tsx
@@ -1,24 +1,7 @@
-import { useEffect } from "react";
-import sweatSmartLogo from "/sweatsmart-logo.svg";
-import axios from "axios";
 
-async function sendCode(code: string) {
-  const data = await axios.get(
-    `https://sweatsmart-be.vercel.app/strava_auth?code=${code}`
-  );
-  console.log(data["data"]);
-}
+import sweatSmartLogo from "/sweatsmart-logo.svg";
 
 function Home() {
-  const params = new URLSearchParams(document.location.search);
-  const code = params.get("code");
-
-  useEffect(() => {
-    if (code) {
-      sendCode(code);
-    }
-  }, [code]);
-
   return (
     <>
       <div>
@@ -38,7 +21,7 @@ function Home() {
           onClick={async () => {
             window.location.href = `https://www.strava.com/oauth/authorize?client_id=120918&redirect_uri=${
               import.meta.env.VITE_REDIRECT_URI
-            }&response_type=code&approval_prompt=auto&scope=activity:write,activity:read`;
+            }/workouts&response_type=code&approval_prompt=auto&scope=activity:write,activity:read`;
           }}
         ></button>
       </div>
